refactor(infra): move PHC string logic to parser for single responsibility
Pbkdf2PhcString is now a pure data class with internal constructor.
All parsing and construction logic lives in Pbkdf2PhcStringParser.
TryParseParams now rejects unknown parameter keys.
